chore: move fetchAppOpportunities into react-queries (#6449)
* chore: move fetchAppOpportunities into react-queries

* chore: move react-queries into separate files
